Broke out constructors for easier modification.
