+++ frontend/public/js/structures/create_actor.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var Immutable  = require('Immutable'),
+    createNode = require('./create_node');
+
+module.exports = function createActorNode(model, data, gui) {
+    return createNode(model, 'actor', data, gui);
+};
